# Make PooledByteBufAllocator friendly with virtual threads