--- a/buffer/src/main/java/io/netty/buffer/PooledByteBuf.java
+++ b/buffer/src/main/java/io/netty/buffer/PooledByteBuf.java
@@ -18,6 +18,7 @@ package io.netty.buffer;
 
 import io.netty.util.Recycler.EnhancedHandle;
 import io.netty.util.internal.ObjectPool.Handle;
+import io.netty.util.internal.ObjectUtil;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -29,7 +30,7 @@ import java.nio.channels.ScatteringByteChannel;
 
 abstract class PooledByteBuf<T> extends AbstractReferenceCountedByteBuf {
 
-    private final EnhancedHandle<PooledByteBuf<T>> recyclerHandle;
+    private final Handle<PooledByteBuf<T>> recyclerHandle;
 
     protected PoolChunk<T> chunk;
     protected long handle;
@@ -44,7 +45,7 @@ abstract class PooledByteBuf<T> extends AbstractReferenceCountedByteBuf {
     @SuppressWarnings("unchecked")
     protected PooledByteBuf(Handle<? extends PooledByteBuf<T>> recyclerHandle, int maxCapacity) {
         super(maxCapacity);
-        this.recyclerHandle = (EnhancedHandle<PooledByteBuf<T>>) recyclerHandle;
+        this.recyclerHandle = ObjectUtil.checkNotNull((Handle<PooledByteBuf<T>>) recyclerHandle, "recyclerHandle");
     }
 
     void init(PoolChunk<T> chunk, ByteBuffer nioBuffer,
@@ -177,7 +178,12 @@ abstract class PooledByteBuf<T> extends AbstractReferenceCountedByteBuf {
             tmpNioBuf = null;
             chunk = null;
             cache = null;
-            this.recyclerHandle.unguardedRecycle(this);
+            if (recyclerHandle instanceof EnhancedHandle) {
+                EnhancedHandle<PooledByteBuf<T>> enhancedHandle = (EnhancedHandle<PooledByteBuf<T>>) recyclerHandle;
+                enhancedHandle.unguardedRecycle(this);
+            } else {
+                recyclerHandle.recycle(this);
+            }
         }
     }