--- a/buffer/src/main/java/io/netty/buffer/PooledHeapByteBuf.java
+++ b/buffer/src/main/java/io/netty/buffer/PooledHeapByteBuf.java
@@ -40,6 +40,10 @@ class PooledHeapByteBuf extends PooledByteBuf<byte[]> {
         return buf;
     }
 
+    static PooledHeapByteBuf newInstanceNoThreadLocal(Handle<PooledHeapByteBuf> handle) {
+        return new PooledHeapByteBuf(handle, 0);
+    }
+
     PooledHeapByteBuf(Handle<? extends PooledHeapByteBuf> recyclerHandle, int maxCapacity) {
         super(recyclerHandle, maxCapacity);
     }