Motivation:

In short, this PR makes `PooledByteBufAllocator.DEFAULT` friendly with virtual threads.

It's better to make `PooledByteBufAllocator` friendly with virtual threads, there was another PR https://github.com/netty/netty/pull/14607 done similar thing, but it made a global decision on the constructor to instruct the allocator, this PR instruct the allocator whether to use thread-local based on the thread type, as @franz1981 suggested via https://github.com/netty/netty/pull/14607#discussion_r1894581153.

In this PR, the `PooledByteBufAllocator` use thread-local when the thread type is event-loop(`FastThreadLocalThread`), otherwise, if virtual threads or normal platform threads, the allocator does not use thread-local.

It might be better to let normal platform threads use thread-local too, but it requires a method handler(call `isVirtual()`), I will add this handler if needed in the future. 

Following are some benchmarking results, the benchmark code: [ByteBufAllocatorConcurrentBenchmark](https://github.com/laosijikaichele/netty/blob/tmp/microbench/0301/ByteBufAllocatorConcurrentBenchmark.java), which benchmark on the default `PooledByteBufAllocator.DEFAULT` and `AdaptiveByteBufAllocator`(for comparation), with JDK-21, M1 chip, 8 cores, 16G RAM:

**Use event-loop(`FastThreadLocalThread`) threads(8 threads):**
Before PR:
```
Benchmark                                                    (size)   Mode  Cnt         Score          Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00064  thrpt   20  88330081.136 ±  5037031.882  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00256  thrpt   20  89871472.042 ± 21346695.226  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   01024  thrpt   20  78504680.692 ± 21982671.256  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   04096  thrpt   20  72127653.154 ±  1417842.181  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00064  thrpt   20  91054515.118 ±   419255.100  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00256  thrpt   20  83165530.913 ±  6702270.137  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     01024  thrpt   20  88603270.902 ±   666014.029  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     04096  thrpt   20  79977401.394 ±  8353521.171  ops/s
```
After PR:
```
Benchmark                                                    (size)   Mode  Cnt         Score         Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00064  thrpt   20  73216063.559 ± 2040971.965  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00256  thrpt   20  76767063.513 ± 1730893.726  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   01024  thrpt   20  99035714.952 ± 8729132.519  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   04096  thrpt   20  66960699.666 ± 7575289.284  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00064  thrpt   20  87781083.219 ±  363728.822  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00256  thrpt   20  90460389.702 ±  360119.721  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     01024  thrpt   20  89990609.228 ± 1368144.074  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     04096  thrpt   20  91623220.419 ± 1618839.974  ops/s
```


**Use normal platform threads(8 threads):**
Before PR:
```
Benchmark                                                    (size)   Mode  Cnt         Score         Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00064  thrpt   20  69719960.924 ± 3544372.539  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00256  thrpt   20  61882436.213 ± 1334615.684  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   01024  thrpt   20  60091641.208 ±  727157.051  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   04096  thrpt   20  60916323.123 ± 3760421.873  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00064  thrpt   20  71549724.503 ±  831271.688  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00256  thrpt   20  71959285.223 ±  530098.030  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     01024  thrpt   20  70358120.108 ±  359661.461  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     04096  thrpt   20  68924374.063 ±  889450.468  ops/s
```
After PR:
```
Benchmark                                                    (size)   Mode  Cnt         Score          Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00064  thrpt   20  77024154.725 ±  6860884.114  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   00256  thrpt   20  75523953.520 ±  2820518.874  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   01024  thrpt   20  70797135.674 ±  3150582.623  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleaseAdaptive   04096  thrpt   20  57482330.057 ± 10720046.853  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00064  thrpt   20  64865343.834 ±   681320.296  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     00256  thrpt   20  62511783.012 ±  1381261.696  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     01024  thrpt   20  62956876.888 ±  1077958.872  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled     04096  thrpt   20  63649065.328 ±  1715389.715  ops/s
```

Now let's benchmark with virtual threads, because JMH does not work well with virtual threads, so I use the custom test code: [PooledByteBufVirtualThreadBenchMark](https://github.com/laosijikaichele/netty/blob/tmp/microbench/0301/PooledByteBufVirtualThreadBenchMark.java).

**Use virtual threads(1000000 threads):**
**(NOTE: The less duration, the better performance)**
Before PR:
```
Adaptive alloc avg duration: 1231
Default alloc avg duration: 1325
---------------------------------------
Adaptive alloc avg duration: 1089
Default alloc avg duration: 1128
---------------------------------------
Adaptive alloc avg duration: 1184
Default alloc avg duration: 1169
---------------------------------------
Adaptive alloc avg duration: 1391
Default alloc avg duration: 1173
---------------------------------------
Adaptive alloc avg duration: 1555
Default alloc avg duration: 1222
---------------------------------------
```
After PR:
```
Adaptive alloc avg duration: 1103
Default alloc avg duration: 892
---------------------------------------
Adaptive alloc avg duration: 1291
Default alloc avg duration: 834
---------------------------------------
Adaptive alloc avg duration: 1199
Default alloc avg duration: 796
---------------------------------------
Adaptive alloc avg duration: 1303
Default alloc avg duration: 946
---------------------------------------
Adaptive alloc avg duration: 1167
Default alloc avg duration: 879
---------------------------------------
```

Analyse:
(The adaptive alloc is mainly used for reference, let's focus on the default alloc)

**When the thread type is event-loop: 
The default alloc performance has no obvious change after PR.**

**When the thread type is normal platform thread: 
The default alloc performance shows degradation after PR: degraded -8% ~ -14%.**
(which makes sense, because it does not use thread-local).

**When the thread type is virtual thread: 
The default alloc performance shows obvious improvement after PR: improved 24+% ~ +48%**

Currently this PR is in draft stage, will update if needed...