@laosijikaichele 2025-03-02
Added virtual thread checker method:
https://github.com/netty/netty/blob/91a9024aa0b4f7f8a136600e4a029e7da3780a91/common/src/main/java/io/netty/util/internal/PlatformDependent0.java#L547-L591

I took some reference from @franz1981 's nice code: :+1:
https://github.com/franz1981/netty/blob/4.1_loom_fix/common/src/main/java/io/netty/util/internal/PlatformDependent0.java#L488-L523