@laosijikaichele 2025-03-03
After added virtual thread checker method, now the logic is: 
If the current thread is confirmed not a virtual thread, then use thread-local, otherwise does not use thread-local.

Following are benchmarking results, the benchmark code: [ByteBufAllocatorConcurrentBenchmark](https://github.com/laosijikaichele/netty/blob/tmp/microbench/0303/ByteBufAllocatorConcurrentBenchmark.java), which benchmark on the default `PooledByteBufAllocator.DEFAULT`, with JDK-21, M1 chip, 8 cores, 16G RAM:

**Use event-loop(FastThreadLocalThread) threads(8 threads):
Before PR:**
```
Benchmark                                                  (size)   Mode  Cnt         Score         Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00064  thrpt   20  77665701.820 ± 1986618.539  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00256  thrpt   20  92789351.995 ± 7825150.897  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   01024  thrpt   20  79836500.864 ±  683605.088  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   04096  thrpt   20  86699914.060 ± 7553827.740  ops/s 
```
**After PR:**
```
Benchmark                                                  (size)   Mode  Cnt         Score         Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00064  thrpt   20  94220144.101 ± 9192095.031  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00256  thrpt   20  82727423.305 ± 1240964.805  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   01024  thrpt   20  94551825.760 ±  568164.136  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   04096  thrpt   20  77878187.635 ± 1119117.338  ops/s
```

**Use normal platform threads(8 threads):
Before PR:**
```
Benchmark                                                  (size)   Mode  Cnt         Score         Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00064  thrpt   20  81281653.328 ± 1379203.806  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00256  thrpt   20  78324116.363 ±  894848.687  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   01024  thrpt   20  63841258.411 ± 8224392.277  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   04096  thrpt   20  72816155.401 ± 2958170.572  ops/s
```
**After PR:**
```
Benchmark                                                  (size)   Mode  Cnt         Score         Error  Units
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00064  thrpt   20  73325028.639 ±  289525.457  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   00256  thrpt   20  73920169.373 ± 2490022.796  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   01024  thrpt   20  70498814.681 ±  696607.576  ops/s
ByteBufAllocatorConcurrentBenchmark.allocateReleasePooled   04096  thrpt   20  68612937.924 ±  761839.482  ops/s
```

Then benchmark with virtual threads by [PooledByteBufVirtualThreadBenchMark](https://github.com/laosijikaichele/netty/blob/tmp/microbench/0301/PooledByteBufVirtualThreadBenchMark.java).
**Use virtual threads(1000000 threads):
(NOTE: The less duration, the better performance)
Before PR:**
```
Adaptive alloc avg duration: 1097
Default alloc avg duration: 969
---------------------------------------
Adaptive alloc avg duration: 1153
Default alloc avg duration: 1065
---------------------------------------
Adaptive alloc avg duration: 1237
Default alloc avg duration: 1149
---------------------------------------
Adaptive alloc avg duration: 1266
Default alloc avg duration: 1000
---------------------------------------
Adaptive alloc avg duration: 1242
Default alloc avg duration: 1039
---------------------------------------
```
**After PR:**
```
Adaptive alloc avg duration: 1072
Default alloc avg duration: 776
---------------------------------------
Adaptive alloc avg duration: 1255
Default alloc avg duration: 769
---------------------------------------
Adaptive alloc avg duration: 1260
Default alloc avg duration: 729
---------------------------------------
Adaptive alloc avg duration: 1306
Default alloc avg duration: 836
---------------------------------------
Adaptive alloc avg duration: 1256
Default alloc avg duration: 870
---------------------------------------
```

Analyse:
Use event-loop threads(`FastThreadLocalThread`)**:
**The default alloc performance shows neutral number after PR: 2/4 degraded(-11%, -10%), 2/4 improved(+10%, +22%).**

Use normal platform threads:
**The default alloc performance slightly tend to degradation after PR: 3/4 degraded (-10%, -6%, -6%), 1/4 improved:(+11%).
(I think the cost mainly due to the thread checker method which calls reflection `Method::invoke`).**

Use virtual threads:
**The default alloc performance shows obvious improvement after PR: improved +20% ~ +57%**