@laosijikaichele 2025-03-04
<img src="https://github.com/user-attachments/assets/1ad913ca-687d-4f12-885c-b60f68e50d8a" width="300px">


- Virtual thread: Do not use thread-local.

- Unknown(Not able to detect the thread type): Do not use thread-local.

- Non-virtual thread(normal platform thread,FastThreadLocalThread): Use thread-local.