@laosijikaichele 2025-04-08
Since 4.1 will only receive bug-fixes, we may not need this improvement, close it for now.
I will open a new PR for 4.2 if still need this.