@franz1981 2025-04-08
Good idea @laosijikaichele but for Netty 4.2: why you don't need it there too?
Adaptive is now a good replacement, from latest tests?