@laosijikaichele 2025-04-08
> I will open a new PR for 4.2 if still need this.

In my opinion, for 4.2, this PR may still be needed, because there may still people will continue to use `PooledByteBufAllocator` in 4.2, if that is the case, I will open a new PR for 4.2. 
WDYT? @franz1981