@franz1981 2025-04-08
To be virtual friendly overall it looks to me a nice thing to have :)
I think we can have some chat with @yawkat about it WDYT?